update: プラグイン用の対応
